backend/requirements.txt
flask==2.3.2
python-dotenv==1.0.1
deepgram-sdk==2.12.0
openai==0.28.1
pydub==0.25.1
